requirements.txt
psutil
pyperclip>=1.8.0
pyinstaller>=6.0.0
opencv-python>=4.8.0
Pillow>=10.0.0
pytesseract>=0.3.10 